+++ server/uptime-kuma-server.js
@@ -110,6 +110,7 @@ class UptimeKumaServer {
         // Set Monitor Types
         UptimeKumaServer.monitorTypeList["real-browser"] = new RealBrowserMonitorType();
         UptimeKumaServer.monitorTypeList["tailscale-ping"] = new TailscalePing();
+        UptimeKumaServer.monitorTypeList["dns"] = new DnsMonitorType();
 
         this.io = new Server(this.httpServer);
     }
@@ -419,3 +420,4 @@ module.exports = {
 // Must be at the end to avoid circular dependencies
 const { RealBrowserMonitorType } = require("./monitor-types/real-browser-monitor-type");
 const { TailscalePing } = require("./monitor-types/tailscale-ping");
+const { DnsMonitorType } = require("./monitor-types/dns");
